refactor(promise-modal): route app-layer imports through the module entry point
