+++ packages/lerx/promise-modal/src/components/Anchor/style.ts
@@ -1,4 +1,4 @@
-import { ModalManager } from '@/promise-modal/app/ModalManager';
+import { ModalManager } from '@/promise-modal/app';
 
 export const anchor = ModalManager.getHashedClassNames('anchor');
